tạo function để thêm số lượng đơn hàng ảo vào db

+++ src/main/java/model/BillDetail.java
@@ -1,11 +1,8 @@
 package model;
-
 import lombok.Builder;
 import lombok.Data;
-
 import java.io.Serializable;
-import java.math.BigInteger;
-
+import java.math.BigDecimal;
 @Data
 @Builder
 public class BillDetail implements Serializable {
@@ -13,7 +10,7 @@ public class BillDetail implements Serializable {
     private int id_bill;
     private int id_product;
     private int quantity;
-    private BigInteger listed_price;
-    private BigInteger current_price;
+    private BigDecimal listed_price;
+    private BigDecimal current_price;
 
 }
